Fix CodeQL scanning alerts: integer overflow, type widening, unsigned diff, and TOCTOU races
Fixes across 19 source files and 1 CodeQL configuration file:

- cpp/integer-multiplication-cast-to-long (22 alerts): Added (LT_SIZE)
  casts before multiplications in LTArrayImpl.c, LTMemory.c,
  ObjectDetectionPostProcessing.c, LTMediaPipelineRingQueueImpl.c,
  LTMediaPipeline.c, MacOSFlashDeviceUnit.c,
  LinuxFlashDeviceUnitClearFile.c, UnitTestLTSystemFS.c

- cpp/comparison-with-wider-type (9 alerts): Changed loop variables
  from u8/u16 to u32/int to match comparison width in
  LTMediaMotionDetection.c, LTMediaMotionDetectionAlgorithm.c,
  RtcpPacket.c, LTSystemCryptoEncoder.c, LTAsset.c, LTShellNet.c,
  UnitTestLTSystemFS.c

- cpp/unsigned-difference-expression-compared-zero (4 alerts):
  Added (s32) casts for signed comparison in
  UnitTestLTMediaMotionDetection.c, ExampleNetRegressTest.c

- cpp/toctou-race-condition (3 alerts): Replaced
  stat-then-fopen/access-then-open/access-then-remove patterns
  with single-operation checks in Arbolate.c,
  LinuxDriverPinsImpl.c, LinuxFlashDeviceUnit.c

- Add .github/codeql/codeql-config.yml to exclude third-party,
  vendor, speex, and ESP32 tool paths from future scans

+++ lt/lt/source/lt/media/objectdetection/ObjectDetectionPostProcessing.c
@@ -903,7 +903,7 @@ void convert_chw_to_hwc_float(float* input, float* output,
     if (!input || !output) return;
 
     if (input == output) {
-        float* temp = lt_malloc(w * h * channels * sizeof(float));
+        float* temp = lt_malloc((size_t)w * h * channels * sizeof(float));
         if (!temp) return;
 
         for (int c = 0; c < channels; c++) {
@@ -914,7 +914,7 @@ void convert_chw_to_hwc_float(float* input, float* output,
                 }
             }
         }
-        lt_memcpy(output, temp, w * h * channels * sizeof(float));
+        lt_memcpy(output, temp, (size_t)w * h * channels * sizeof(float));
         lt_free(temp);
 
     } else {
